+++ CHANGELOG.md
@@ -257,6 +257,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Edit form instead of just adding forms to the project
 - Add and rename forms instead of editing services
 - Notifications only for Project Manager Admin
+- Filter templates by selected forms
 
 ### Removed
 - RStudio Keycloak Groups